Rename SQLQuery1.sql to Game Sales Data Exploration.sql
Game Sales Data Exploration is a repository that contains code and data for exploring the global game sales market. The data includes information on game sales by platform and genre. The code can be used to generate visualizations and insights about the game sales market.

This repository is a valuable resource for anyone interested in learning more about the global game sales market. The data and code can be used to answer questions about the market, such as:

Which platforms are the most popular for game sales?
What genres are the most popular for game sales?